src/app/requirements.txt
aiogram~=3.0.0b3
aiohttp
tortoise-orm[asyncmy]
cryptography